apps/pyproject.toml
[tool.poetry]
name = "sampleinspectionselector"
version = "0.1.0"
description = ""
authors = ["Jun Izumida <jun.izumida@agc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
celery = "^5.4.0"
redis = "^5.0.8"
psycopg2-binary = "^2.9.9"
pymongo = "^4.8.0"
pyyaml = "^6.0.2"
django-dotenv = "^1.4.2"
django-environ = "^0.11.2"
django-debug-toolbar = "^4.4.6"
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"
django-cors-headers = "^4.4.0"
strawberry-graphql-django = "^0.47.1"
strawberry-graphql = "^0.237.3"
pysmb = "^1.2.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
